Factory is not spewing out hot red messages now. This is good

+++ src/vooga/rts/gamedesign/factories/UnitDecoder.java
@@ -52,7 +52,7 @@ public class UnitDecoder extends Decoder {
 			String img = nElement.getElementsByTagName(IMAGE_TAG).item(0).getTextContent();
 			String sound = nElement.getElementsByTagName(SOUND_TAG).item(0).getTextContent();
 			int health = Integer.parseInt(nElement.getElementsByTagName(HEALTH_TAG).item(0).getTextContent());
-			int buildTime = Integer.parseInt(nElement.getElementsByTagName(TIME_TAG).item(0).getTextContent());
+			double buildTime = Double.parseDouble(nElement.getElementsByTagName(TIME_TAG).item(0).getTextContent());
 			
 			Unit unit = (Unit) ReflectionHelper.makeInstance(headClass, new Pixmap(img), 
 																		new Location3D(0,0,0),
